Create separate tests for plugins
